introduce logical attributes 4

+++ VeySicK.h
@@ -359,11 +359,11 @@ extern std::shared_ptr<VskStrMemoryBlock> vsk_str_block;
 
 struct VskLogAttr
 {
-    uint8_t m_palette : 7;      // 0～7: black, blue, ..., yellow, white
-    uint8_t m_effect : 7;       // 0～7: secret, blink, and/or reverse
-    uint8_t m_semigra : 1;      // 0 or 1
-    uint8_t m_upperline : 1;    // 0 or 1
-    uint8_t m_underline : 1;    // 0 or 1
+    uint32_t m_palette : 7;      // 0～7: black, blue, ..., yellow, white
+    uint32_t m_effect : 7;       // 0～7: secret, blink, and/or reverse
+    uint32_t m_semigra : 1;      // 0 or 1
+    uint32_t m_upperline : 1;    // 0 or 1
+    uint32_t m_underline : 1;    // 0 or 1
 
     void reset()
     {
